Fix an error in the tag finding logic.
Basically, since we've set the search positions at starts and ends
of reads during the actual search, we don't need to verify those
positions after making a match.  Also, given the way that we've
sliced, the base positions are relative to the slice, not the
sequence, which makes things a PITA to deal with.